about me stylee css
Styling the about me section. Not quite finished yet. I want bullet points on the lists but it has also added them to the nav bar which I dont want. potentially use a class on the indivual li elements in the ul of the about me section to navigate this. or in the nav bar section. its looking good and will continue to improve