style.css
added a border to be able to see the changes on the webpage when I made them Was curious as to why the profile section wasnt taking up 100vh. I had the height set to 50 vh. Changing to 80 has fixed the problem. Need to add the arrows to change sections.