updates
changed the titles of nav bar links to more appropriate descriptions of the content im going to have on the webpage. changed href to match id of about me section and is working as intended. added arrow images to both sections. changed profile css to position relative and this solved the issue of the arrow not being placed correctly and shifting all the content over on the page. also styled the arrow icon itself.